Update truncate.js
This is one of numerous ways to truncate a string.

Three variable are declared. (1. ) Variable a which stores result of the three dots strings subtracted from the provided num because it all add to he length of the string. (2. ) Variable sliceStr which stores the truncated string after the slice method has acted on it and also concatenate the dots. (3. ) Variable strLen which stores the return value of the length of the string provided.

Then follows our condition statement which checks the following conditions:
i. strLen is LESS THAN the length of the three dots OR num is LESS THAN the length of the three dots then we slice with the number provided.
ii. num  GREATER THAN the length strLen, no need of truncating, we just return the provided string.
iii. else we subtract 3 from the provided number will is stored in variable a declared earlier.